fix(types): mypy auf 0 Fehler (483 → 0) — drei echte Bugs gefunden
Bugfixes (durch mypy aufgedeckt):
- pipeline_registry: audio_manager.stop() lief nie (async ohne await)
- research/orchestrator: rank_urls_by_relevance ohne Pflichtarg
  llm_history → TypeError im Forced-Research-Ranking-Pfad
- process_utils: Import des nie existierenden Modul-Attributs
  vllm_manager → vLLM-Stop lief nie; Instanz kommt jetzt aus dem
  Backend-State

Struktur:
- mypy.ini: Reflex-Strukturcodes zentral für ui/* (Var-Indexing,
  Event-Pseudo-Calls) und state/* (nur attr-defined/has-type,
  Cross-Mixin) — arg-type/assignment bleiben überall scharf
- mypy.ini: google-Namespace-Packages mit follow_imports=skip —
  behebt den inkrementellen Cache-Crash ('Cannot find module google')
- AudioOutputChannel: Local/Browser erben die Protocol-Defaults;
  get_stream_start_offset als dritter Default im Protocol
- TTS-Basis: calibration_vram_reserve_mb als Property (Engines
  overriden mit @property; xtts-Redundanz raus)

Dazu: Telegram-None-Guards (Channel-Posts ohne User), IMAP bytes/str,
except-/Loop-Variablen-Shadowing, EventSpec-Rückgabetypen der Reflex-
Handler, AsyncGenerator-Signaturen, gezielte Casts statt Any-Returns.

ruff sauber, mypy 0/238 Dateien, 447/447 Tests grün.

+++ aifred/lib/gpu_utils.py
@@ -68,8 +68,8 @@ def total_actual_vram_gb(gpu_info) -> float:
     Single source of truth for VRAM total display.
     """
     if gpu_info.all_gpu_vram_mb:
-        return round(sum(gpu_info.all_gpu_vram_mb) / 1024, 1)
-    return round(gpu_info.vram_mb / 1024, 1)
+        return float(round(sum(gpu_info.all_gpu_vram_mb) / 1024, 1))
+    return float(round(gpu_info.vram_mb / 1024, 1))
 
 
 # NOTE: is_model_loaded() was moved to backends/base.py as abstractmethod
